:man: :video_game: AQ verification can be performed
By scraping user pages, a script can be run to verify a user's claimed
AQ account. Currently this is done by converting some of their ID to a
hexadecimal and seeing if the AQ account has their armor trim set to
that same hexadecimal color.
This can be set up to run as a cron job or right when they try to set
their AQ name too.